Create blank __main__ files
